+++ share/dotfiles/.config/hypr/scripts/gtk.sh
@@ -16,20 +16,27 @@ icon_theme="$(grep 'gtk-icon-theme-name' "$config" | sed 's/.*\s*=\s*//')"
 cursor_theme="$(grep 'gtk-cursor-theme-name' "$config" | sed 's/.*\s*=\s*//')"
 cursor_size="$(grep 'gtk-cursor-theme-size' "$config" | sed 's/.*\s*=\s*//')"
 font_name="$(grep 'gtk-font-name' "$config" | sed 's/.*\s*=\s*//')"
+prefer_dark_theme="$(grep 'gtk-application-prefer-dark-theme' "$config" | sed 's/.*\s*=\s*//')"
 terminal=$(cat $HOME/.config/ml4w/settings/terminal.sh)
 
-echo $gtk_theme
-echo $icon_theme
-echo $cursor_theme
-echo $cursor_size
-echo $font_name
-echo $terminal
+echo "GTK-Theme:" $gtk_theme
+echo "Icon Theme:" $icon_theme
+echo "Cursor Theme:" $cursor_theme
+echo "Cursor Size:" $cursor_size
+if [ $prefer_dark_theme == "0" ]; then
+    prefer_dark_theme_value="prefer-light"
+else
+    prefer_dark_theme_value="prefer-dark"
+fi
+echo "Color Theme:" $prefer_dark_theme_value
+echo "Font Name:" $font_name
+echo "Terminal:" $terminal
 
 gsettings set "$gnome_schema" gtk-theme "$gtk_theme"
 gsettings set "$gnome_schema" icon-theme "$icon_theme"
 gsettings set "$gnome_schema" cursor-theme "$cursor_theme"
 gsettings set "$gnome_schema" font-name "$font_name"
-gsettings set "$gnome_schema" color-scheme "prefer-dark"
+gsettings set "$gnome_schema" color-scheme "$prefer_dark_theme_value"
 
 gsettings set com.github.stunkymonkey.nautilus-open-any-terminal terminal "$terminal"
 gsettings set com.github.stunkymonkey.nautilus-open-any-terminal use-generic-terminal-name "true"