+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 2.9.8
 --------------------------------------------------------
 - All ML4W Apps (except Hyprland Settings for now) are now flatpak apps.
 - New ML4W Sidebar App give quick access to selected configurations. It replaces the ags sidebar.
+- All apps are now supporting light theme. Change it in nwg-look for GTK or QT6.
 - SwayNC added as new notification app and center with additional features.
 - kitty, nvim, .bashrc and .zshrc requires user confirmation for overwriting.
 - All required additional downloads required for the installation will be stored in .ml4w folder